+++ web/user/src/locales/ja_jp.json
@@ -103,7 +103,6 @@
       "viewMoreText": "もっと見る",
       "pastMarcheLinkText": "過去のライブ配信はこちら",
       "productsLinkText": "購入できる商品一覧はこちら",
-      "procut": "購入できる商品一覧",
       "archiveListSubTitle": "アーカイブ一覧",
       "productListSubTitle": "商品動画一覧",
       "experienceListSubTitle": "体験動画一覧",
